dpi600_phase1-2-3_complete/ai-training/requirements.txt
# DPI-600 Drug Logo Recognition - Requirements
# =============================================

# Core ML Framework
torch>=2.0.0
torchvision>=0.15.0
torchaudio>=2.0.0

# Image Processing
Pillow>=10.0.0
opencv-python>=4.8.0
albumentations>=1.3.1

# Data Processing
numpy>=1.24.0
pandas>=2.0.0
scikit-learn>=1.3.0

# Visualization
matplotlib>=3.7.0
seaborn>=0.12.0

# Model Export
onnx>=1.14.0
onnxruntime>=1.15.0

# MLOps
mlflow>=2.6.0

# Azure ML (optional - for cloud training)
# azureml-core>=1.53.0
# azureml-mlflow>=1.53.0

# API Server (optional)
flask>=2.3.0
gunicorn>=21.0.0

# Utilities
tqdm>=4.65.0
pyyaml>=6.0.0
python-dotenv>=1.0.0

# Testing
pytest>=7.4.0
